app/requirements.txt
flask
python-dotenv
requests
urllib